+++ src/features/XIT/ACT/actions/mtra/config.ts
@@ -1,0 +1,6 @@
+export const configuredLocation = 'Configure on Execution';
+
+export interface Config {
+  origin?: string;
+  destination?: string;
+}